+++ src/main/java/dev/qixils/fahare/Fahare.java
@@ -17,16 +17,20 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
+import org.bukkit.event.entity.EntityMountEvent;
 import org.bukkit.event.entity.EntityPortalEvent;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerPortalEvent;
 import org.bukkit.event.player.PlayerRespawnEvent;
 import org.bukkit.event.player.PlayerTeleportEvent;
+import org.bukkit.persistence.PersistentDataType;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.LocalDate;
@@ -43,16 +47,22 @@ public final class Fahare extends JavaPlugin implements Listener {
     private static final Random RANDOM = new Random();
     private final NamespacedKey fakeOverworldKey = new NamespacedKey(this, "overworld");
     private final NamespacedKey limboWorldKey = new NamespacedKey(this, "limbo");
+    private final NamespacedKey lastResetIdKey = new NamespacedKey(this, "last_reset_id");
     private final Map<UUID, Integer> deaths = new HashMap<>();
+    private final Set<UUID> playersPendingMountCancel = new HashSet<>();
+    private int currentResetId = 0;
     private World limboWorld;
     private Path worldContainer;
     private @Nullable Path backupContainer;
+    private @Nullable Path resetIdPath;
     private boolean resetting = false;
     // config
     private boolean backup = true;
     private boolean autoReset = true;
     private boolean anyDeath = false;
     private int lives = 1;
+    private Difficulty forceDifficulty = null;
+    private long forceSeed = 0L;
 
     private static @NotNull World overworld() {
         return Objects.requireNonNull(Bukkit.getWorld(REAL_OVERWORLD_KEY), "Overworld not found");
@@ -62,20 +72,61 @@ public final class Fahare extends JavaPlugin implements Listener {
         return Objects.requireNonNullElseGet(Bukkit.getWorld(fakeOverworldKey), this::createFakeOverworld);
     }
 
+    private long getNewSeed() {
+        if (forceSeed != 0L) return forceSeed;
+        return RANDOM.nextLong();
+    }
+
+    private Difficulty getNewDifficulty() {
+        if (forceDifficulty != null) return forceDifficulty;
+        return overworld().getDifficulty();
+    }
+
+    private boolean getNewHardcore(Difficulty difficulty) {
+        return autoReset && lives <= 1 && difficulty == Difficulty.HARD;
+    }
+
     private @NotNull World createFakeOverworld() {
         // Create fake overworld
-        long seed = RANDOM.nextLong();
+        Difficulty difficulty = getNewDifficulty();
+        long seed = getNewSeed();
         getComponentLogger().info(translatable("fhr.log.overworld-seed", text(seed)));
-        WorldCreator creator = new WorldCreator(fakeOverworldKey).copy(overworld()).seed(seed);
+        WorldCreator creator = new WorldCreator(fakeOverworldKey)
+                .copy(overworld())
+                .seed(seed)
+                .hardcore(getNewHardcore(difficulty));
+
         World world = Objects.requireNonNull(creator.createWorld(), "Could not load fake overworld");
-        world.setDifficulty(overworld().getDifficulty());
+        world.setDifficulty(difficulty);
+
         return world;
+    }
+
+    private void saveResetId() {
+        // todo: if we have to expand any further than 1 saved value then we should get a json but for now this is fine
+        if (resetIdPath == null) return;
+
+        try {
+            Files.createDirectories(resetIdPath.getParent());
+            Files.writeString(resetIdPath, String.valueOf(currentResetId));
+        } catch (IOException e) {
+            getComponentLogger().warn(translatable("fhr.log.error.save-reset-id"), e);
+        }
     }
 
     @Override
     public void onEnable() {
         // Load config
         loadConfig();
+        resetIdPath = getDataPath().resolve("current-reset-id.txt");
+        try {
+            if (Files.exists(resetIdPath)) {
+                String idString = Files.readString(resetIdPath).trim();
+                currentResetId = Integer.parseInt(idString);
+            }
+        } catch (Exception ignored) {
+            // file probably doesn't exist yet
+        }
 
         // Create backup folder
         worldContainer = Bukkit.getWorldContainer().toPath();
@@ -87,6 +138,33 @@ public final class Fahare extends JavaPlugin implements Listener {
             } catch (Exception e) {
                 getComponentLogger().error(translatable("fhr.log.error.backup-folder"), e);
                 backupContainer = null;
+            }
+        }
+
+        // Load server properties to extract level-seed
+        Properties properties = new Properties();
+        Path serverProperties = worldContainer.resolve("server.properties");
+        if (Files.exists(serverProperties)) {
+            try (InputStream contents = Files.newInputStream(serverProperties)) {
+                properties.load(contents);
+                try {
+                    String levelSeed = properties.getProperty("level-seed", "");
+                    forceSeed = Long.parseLong(levelSeed);
+                    getComponentLogger().info(translatable("fhr.log.info.found-seed", text(forceSeed)));
+                } catch (Exception e) {
+                    forceSeed = 0;
+                    getComponentLogger().info(translatable("fhr.log.info.missing-seed"));
+                }
+
+                String difficulty = properties.getProperty("difficulty", "");
+                try {
+                    forceDifficulty = Difficulty.valueOf(difficulty.toUpperCase(Locale.US));
+                    getComponentLogger().warn(translatable("fhr.log.info.difficulty", text(String.valueOf(forceDifficulty))));
+                } catch (Exception e) {
+                    getComponentLogger().warn(translatable("fhr.log.error.difficulty", text(difficulty)));
+                }
+            } catch (Exception e) {
+                getComponentLogger().warn(translatable("fhr.log.error.properties"), e);
             }
         }
 
@@ -142,7 +220,7 @@ public final class Fahare extends JavaPlugin implements Listener {
 
         // Register events and tasks
         Bukkit.getPluginManager().registerEvents(this, this);
-        Bukkit.getScheduler().runTaskTimer(this, () -> {
+        Bukkit.getGlobalRegionScheduler().runAtFixedRate(this, $ -> {
             // Teleport players from real overworld
             Location destination = fakeOverworld.getSpawnLocation();
             for (Player player : overworld().getPlayers()) {
@@ -177,6 +255,26 @@ public final class Fahare extends JavaPlugin implements Listener {
         return !isDead(player);
     }
 
+    private int getPlayerLastResetId(Player player) {
+        return player.getPersistentDataContainer().getOrDefault(lastResetIdKey, PersistentDataType.INTEGER, 0);
+    }
+
+    private void setPlayerLastResetId(Player player, int resetId) {
+        player.getPersistentDataContainer().set(lastResetIdKey, PersistentDataType.INTEGER, resetId);
+    }
+
+    private void resetPlayer(Player player) {
+        player.setGameMode(GameMode.SURVIVAL);
+        player.getInventory().clear();
+        player.getEnderChest().clear();
+        player.setLevel(0);
+        player.setExp(0);
+        player.setHealth(20);
+        player.setFoodLevel(20);
+        player.setSaturation(5);
+        setPlayerLastResetId(player, currentResetId);
+    }
+
     private void deleteNextWorld(List<World> worlds, @Nullable Path backupDestination) {
         // check if all worlds are deleted
         if (worlds.isEmpty()) {
@@ -192,7 +290,7 @@ public final class Fahare extends JavaPlugin implements Listener {
 
         // check if worlds are ticking
         if (Bukkit.isTickingWorlds()) {
-            Bukkit.getScheduler().runTaskLater(this, () -> deleteNextWorld(worlds, backupDestination), 1);
+            Bukkit.getGlobalRegionScheduler().runDelayed(this, $ -> deleteNextWorld(worlds, backupDestination), 1);
             return;
         }
 
@@ -201,9 +299,15 @@ public final class Fahare extends JavaPlugin implements Listener {
         String worldName = world.getName();
         Component worldKey = text(worldName);
         WorldCreator creator = new WorldCreator(worldName, world.getKey());
-        long seed = RANDOM.nextLong();
+
+        Difficulty difficulty = getNewDifficulty();
+        long seed = getNewSeed();
         getComponentLogger().info(translatable("fhr.log.seed", worldKey, text(seed)));
-        creator.copy(world).seed(seed);
+
+        creator
+                .copy(world)
+                .seed(seed)
+                .hardcore(getNewHardcore(difficulty));
 
         // unload world
         if (Bukkit.unloadWorld(world, backup)) {
@@ -221,7 +325,11 @@ public final class Fahare extends JavaPlugin implements Listener {
                 }
 
                 // create new world
-                creator.createWorld();
+                World newWorld = creator.createWorld();
+                if (newWorld == null) throw new IllegalStateException("World was null");
+
+                world.setDifficulty(difficulty);
+
                 Bukkit.getServer().sendMessage(translatable("fhr.chat.success", worldKey));
             } catch (Exception e) {
                 Component error = translatable("fhr.chat.error", NamedTextColor.RED, worldKey);
@@ -232,7 +340,7 @@ public final class Fahare extends JavaPlugin implements Listener {
             Bukkit.getServer().sendMessage(translatable("fhr.chat.error", NamedTextColor.RED, worldKey));
         }
 
-        Bukkit.getScheduler().runTaskLater(this, () -> deleteNextWorld(worlds, backupDestination), 1);
+        Bukkit.getGlobalRegionScheduler().runDelayed(this, $ -> deleteNextWorld(worlds, backupDestination), 1);
     }
 
     public synchronized void reset() {
@@ -241,22 +349,21 @@ public final class Fahare extends JavaPlugin implements Listener {
         if (limboWorld == null)
             return;
         deaths.clear();
+
+        // Update and save current reset ID to keep track of players that were offline during the reset
+        currentResetId++;
+        saveResetId();
+
         // teleport all players to limbo
         Location destination = new Location(limboWorld, 0, 100, 0);
         for (Player player : Bukkit.getOnlinePlayers()) {
+            resetPlayer(player);
             player.setGameMode(GameMode.SPECTATOR);
-            player.getInventory().clear();
-            player.getEnderChest().clear();
-            player.setLevel(0);
-            player.setExp(0);
             player.teleport(destination);
-            player.setHealth(20);
-            player.setFoodLevel(20);
-            player.setSaturation(5);
         }
         // check if worlds are ticking
         if (Bukkit.isTickingWorlds()) {
-            Bukkit.getScheduler().runTaskLater(this, this::reset, 1);
+            Bukkit.getGlobalRegionScheduler().runDelayed(this, $ -> reset(), 1);
             return;
         }
         resetting = true;
@@ -306,7 +413,7 @@ public final class Fahare extends JavaPlugin implements Listener {
         addDeathTo(player.getUniqueId());
         if (isAlive(player.getUniqueId()))
             return;
-        Bukkit.getScheduler().runTaskLater(this, () -> {
+        Bukkit.getRegionScheduler().runDelayed(this, player.getLocation(), $ -> {
             player.setGameMode(GameMode.SPECTATOR);
             player.spigot().respawn();
             resetCheck(true);
@@ -347,8 +454,35 @@ public final class Fahare extends JavaPlugin implements Listener {
     @EventHandler(priority = EventPriority.LOWEST)
     public void onPlayerJoin(PlayerJoinEvent event) {
         Player player = event.getPlayer();
-        if (player.getWorld().getKey().equals(REAL_OVERWORLD_KEY))
+
+        // Check if player missed a reset
+        int playerLastReset = getPlayerLastResetId(player);
+        if (playerLastReset < currentResetId) {
+            // Player was offline during reset, apply reset procedure
+            resetPlayer(player);
             player.teleport(fakeOverworld().getSpawnLocation());
+
+            // Store the player's UID for one tick to stop the potential EntityMountEvent
+            playersPendingMountCancel.add(player.getUniqueId());
+            Bukkit.getGlobalRegionScheduler().runDelayed(
+                    this,
+                    $ -> playersPendingMountCancel.remove(player.getUniqueId()),
+                    1L
+            );
+        } else if (player.getWorld().getKey().equals(REAL_OVERWORLD_KEY)) {
+            // Normal case: redirect from real overworld to fake overworld
+            player.teleport(fakeOverworld().getSpawnLocation());
+        }
+    }
+
+    @EventHandler(priority = EventPriority.LOWEST)
+    public void onEntityMount(EntityMountEvent event) {
+        if (playersPendingMountCancel.contains(event.getEntity().getUniqueId())) {
+            // Player is re-mounting a previously mounted entity from a reset world, cancel and remove it
+            // This causes a vanilla log message "Couldn't reattach entity to player", but it is harmless
+            event.setCancelled(true);
+            event.getMount().remove();
+        }
     }
 
     @EventHandler(priority = EventPriority.LOWEST)